YEL-139 [test] API url 변경에 따른 테스트 코드 수정 및 일부 로직 개선
