+++ src/test/java/com/yello/server/domain/authorization/FakeConnectionManager.java
@@ -1,0 +1,13 @@
+package com.yello.server.domain.authorization;
+
+import com.yello.server.domain.authorization.dto.kakao.KakaoTokenInfo;
+import com.yello.server.global.common.manager.ConnectionManager;
+import org.springframework.http.ResponseEntity;
+
+public class FakeConnectionManager implements ConnectionManager {
+
+    @Override
+    public ResponseEntity<KakaoTokenInfo> getKakaoTokenInfo(String kakaoAccessToken) {
+        return null;
+    }
+}
